backend/requirements.txt
# Django Core
Django>=5.0
djangorestframework
django-cors-headers
python-dotenv
dj-database-url
psycopg[binary]

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Image Processing
Pillow

# Security & Validation
python-magic-bin
pyclamd
bleach

# Authentication
djangorestframework-simplejwt
django-ratelimit

# AI
google-generativeai

# File Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-magic>=0.4.27

# Telegram
telethon>=1.36.0
cryptography>=41.0.0
